fix(one): param in path not handled properly
